odvojio interfejse od helper funkcija
