+++ main.js
@@ -13,7 +13,7 @@ app.use(cors())
 app.use(bodyParser.urlencoded({ extended: false }))
 app.use(bodyParser.json())
 app.use((req, res, next) => {
-  console.log(req.url)
+  console.log(req.method, req.url)
   next()
 })
