+++ utils/helpers.js
@@ -6,30 +6,7 @@ const emailCheck = email => {
   return regex.test(email)
 }
 
-class Response {
-  constructor(message, data = null) {
-    this.message = message
-    this.data = data
-  }
-}
-
-class ErrRes extends Response {
-  constructor(message, data) {
-    super(message, data)
-    this.status = "error"
-  }
-}
-
-class SuccRes extends Response {
-  constructor(message, data) {
-    super(message, data)
-    this.status = "success"
-  }
-}
-
 module.exports = {
   nevalidnaLokacija,
   emailCheck,
-  ErrRes,
-  SuccRes
 }
